fixed some errors

+++ src/main/java/com/icesi/taller1/service/PersonService.java
@@ -63,7 +63,7 @@ public class PersonService {
 	}
 	
 	@Transactional
-	public Iterable<Person> findAllById(long id) {
+	public Iterable<Person> findAllById(Integer id) {
 		ArrayList<Person> persons = (ArrayList<Person>) personDAO.findAll();
 		return persons;
 	}
